# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: News
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://localhost'
paths:
    /api/articles:
        get:
            summary: 'Get Articles'
            description: 'List all the articles'
            parameters:
                -
                    in: query
                    name: page
                    description: 'Which page to show.'
                    example: 10
                    required: false
                    schema:
                        type: integer
                        description: 'Which page to show.'
                        example: 10
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: []
                                    links: { first: 'http://localhost/api/articles?page=1', last: 'http://localhost/api/articles?page=2', prev: 'http://localhost/api/articles?page=9', next: null }
                                    meta: { current_page: 10, from: null, last_page: 2, links: [{ url: 'http://localhost/api/articles?page=9', label: '&laquo; Previous', active: false }, { url: 'http://localhost/api/articles?page=1', label: 1, active: false }, { url: 'http://localhost/api/articles?page=2', label: 2, active: false }, { url: null, label: 'Next &raquo;', active: false }], path: 'http://localhost/api/articles', per_page: 6, to: null, total: 12 }
                                properties:
                                    data: { type: array, example: [] }
                                    links: { type: object, example: { first: 'http://localhost/api/articles?page=1', last: 'http://localhost/api/articles?page=2', prev: 'http://localhost/api/articles?page=9', next: null } }
                                    meta: { type: object, example: { current_page: 10, from: null, last_page: 2, links: [{ url: 'http://localhost/api/articles?page=9', label: '&laquo; Previous', active: false }, { url: 'http://localhost/api/articles?page=1', label: 1, active: false }, { url: 'http://localhost/api/articles?page=2', label: 2, active: false }, { url: null, label: 'Next &raquo;', active: false }], path: 'http://localhost/api/articles', per_page: 6, to: null, total: 12 } }
            tags:
                - Articles
            security: []
        post:
            summary: ''
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Articles
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                title:
                                    type: string
                                    description: ''
                                    example: tempore
                                content:
                                    type: string
                                    description: ''
                                    example: modi
                                image:
                                    type: string
                                    format: binary
                                    description: 'The value must be an image.'
                            required:
                                - title
                                - content
                                - image
            security: []
        delete:
            summary: ''
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Articles
            security: []
...